% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLadiesChapters.R
\docType{data}
\name{RLadiesChapters}
\alias{RLadiesChapters}
\title{RLadiesDataset}
\format{
A tibble with 160 rows and 9 variables:
\describe{
\item{chapter}{chr The name of the chapter}
\item{date}{chr The date the Chapter was created}
\item{members}{int The number of members for that Chapter}
\item{lat}{num The latitiude coordinate for the Chapter location}
\item{lon}{num The longitude coordinate for the Chapter location}
\item{city}{chr The name of the city for the Chapter}
\item{country}{chr The country abbreviated name}
}
}
\source{
\url{RLadies Freiburg (2020)}
}
\usage{
RLadiesChapters
}
\description{
RLadies Chapters basic information
}
\keyword{datasets}
