% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLadiesData-package.R
\docType{package}
\name{RLadiesData-package}
\alias{RLadiesData}
\alias{RLadiesData-package}
\title{RLadiesData: RLadiesData}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Zane Dax \email{neptunium0237@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
